+++ src/main.cpp
@@ -37,7 +37,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.4.52"};
+constexpr auto version = std::string_view{"0.4.53"};
 
 struct BootstrapConfigResult final
 {